refactor: editor to a hook
